Implemented a new Notebook class
I implemented a new notebook class, wxAUINotebook, and, as a result, you can now remove tabs and shuffle them around in the tab bar by drag and drop selection. 

I also removed the add and remove tab buttons, and place the "add tab" functionality to the file menu, where the "about" menu item used to be.

Lastly, I commented some of the methods in the tab module.